+++ src/v/cloud_topics/frontend/frontend.cc
@@ -12,8 +12,8 @@
 #include "cloud_storage/types.h"
 #include "cloud_topics/data_plane_api.h"
 #include "cloud_topics/frontend/errc.h"
+#include "cloud_topics/frontend/level_zero_log_reader.h"
 #include "cloud_topics/level_zero/common/extent_meta.h"
-#include "cloud_topics/level_zero/reader/reader.h"
 #include "cloud_topics/level_zero/stm/placeholder.h"
 #include "cloud_topics/logger.h"
 #include "cluster/partition.h"
