+++ src/v/cloud_topics/frontend/level_zero_log_reader.h
